Refactored file structure. removed duplicate files. import declarations written.
